+++ app/src/main/java/com/github/se/assocify/model/entities/Task.kt
@@ -1,0 +1,6 @@
+package com.github.se.assocify.model.entities
+
+data class Task(
+    val name: String,
+    val description: String,
+    val isCompleted: Boolean = false)
