+++ app/src/main/java/com/github/se/assocify/ui/screens/event/map/EventMapScreen.kt
@@ -1,9 +1,9 @@
-package com.github.se.assocify.ui.screens.event.task
+package com.github.se.assocify.ui.screens.event.map
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 
 @Composable
-fun EventScheduleScreen() {
+fun EventMapScreen() {
   Text("This functionality has yet to be implemented")
 }
