+++ app/src/main/java/com/github/se/assocify/ui/screens/event/schedule/EventScheduleScreen.kt
@@ -1,9 +1,9 @@
-package com.github.se.assocify.ui.screens.event.task
+package com.github.se.assocify.ui.screens.event.schedule
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 
 @Composable
-fun EventMapScreen() {
+fun EventScheduleScreen() {
   Text("This functionality has yet to be implemented")
 }
