+++ app/src/main/java/com/github/se/assocify/ui/screens/event/task/EventTaskScreen.kt
@@ -2,8 +2,9 @@ package com.github.se.assocify.ui.screens.event.task
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import com.github.se.assocify.model.entities.Task
 
 @Composable
-fun EventTaskScreen() {
-  Text("This functionality has yet to be implemented")
+fun EventTaskScreen(tasks: List<Task>) {
+
 }
